requirements.txt
ultralytics
opencv-python
numpy
python-dotenv
sqlalchemy
psycopg2-binary